+++ vehicle/src/Vehicle/Verify/QueryFormat/Interface.hs
@@ -3,9 +3,17 @@ module Vehicle.Verify.QueryFormat.Interface where
 import Control.Monad.Except (MonadError)
 import Data.Text (Text)
 import Vehicle.Compile.Error (CompileError)
-import Vehicle.Compile.Prelude (ExternalOutputFormat, MonadLogger)
+import Vehicle.Compile.Prelude (Doc, ExternalOutputFormat, MonadLogger)
 import Vehicle.Verify.Core
 import Vehicle.Verify.QueryFormat.Core
+import Vehicle.Verify.Variable
+
+type FormatQuery =
+  forall m.
+  (MonadLogger m, MonadError CompileError m) =>
+  QueryAddress ->
+  QueryContents ->
+  m Text
 
 -- | A format for an output query that verifiers can parse.
 data QueryFormat = QueryFormat
@@ -13,10 +21,6 @@ data QueryFormat = QueryFormat
     queryOutputFormat :: ExternalOutputFormat,
     supportsStrictInequalities :: Bool,
     -- | The command to format an individual query
-    formatQuery ::
-      forall m.
-      (MonadLogger m, MonadError CompileError m) =>
-      QueryAddress ->
-      QueryContents ->
-      m Text
+    formatQuery :: FormatQuery,
+    compileVar :: forall a. NetworkRationalVariable -> Doc a
   }