+++ vehicle/src/Vehicle/Verify/Verifier.hs
@@ -12,7 +12,9 @@ where
 
 import Vehicle.Verify.Verifier.Core
 import Vehicle.Verify.Verifier.Marabou (marabouVerifier)
+import Vehicle.Verify.Verifier.Test (testVerifier)
 
 verifiers :: VerifierID -> Verifier
 verifiers = \case
   Marabou -> marabouVerifier
+  TestVerifier -> testVerifier